requirements.txt
blis
flask-cors

Flask
scikit-learn
pdfminer.six
python-docx
sentence-transformers
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
spacy==3.7.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
